+++ BuildPiecesCustomized.cs
@@ -25,7 +25,7 @@ namespace BuildPiecesCustomized
     {
         public const string pluginID = "shudnal.BuildPiecesCustomized";
         public const string pluginName = "Build Pieces Customized";
-        public const string pluginVersion = "1.2.2";
+        public const string pluginVersion = "1.2.3";
 
         private readonly Harmony harmony = new Harmony(pluginID);